+++ tortoise/backends/sqlite/executor.py
@@ -57,7 +57,7 @@ class SqliteExecutor(BaseExecutor):
         fields.DatetimeField: to_db_datetime,
     }
     EXPLAIN_PREFIX = "EXPLAIN QUERY PLAN"
-    DB_NATIVE = {bytes, str, int, bool, float}
+    DB_NATIVE = {bytes, str, int, float}
 
     def Parameter(self, pos: int) -> Parameter:
         return Parameter("?")