+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 * Default ``values()`` & ``values_list()`` now includes annotations.
 * Annotations over joins now work correctly with ``values()`` & ``values_list()``
 * Ensure ``GROUP BY`` precedes ``HAVING`` to ensure that filtering by aggregates work correctly.
+* Cast ``BooleanField`` values correctly on SQLite & MySQL
 
 0.15.19
 -------